moved contact force calculation functions from contacts_forces.py to _contact_functions.py to make it less croweded
